don't copy workspaces.yml if already exists

+++ setup.sh
@@ -5,6 +5,8 @@ echo Copying scripts...
 sudo mkdir -p /usr/share/workspaces
 sudo cp workspaces.py workspaces.sh /usr/share/workspaces/
 sudo cp workspaces.completion.sh /etc/bash_completion.d/
-sudo cp workspaces.yml /etc/
+if [ ! -e "/etc/workspaces.yml" ]; then
+    sudo cp workspaces.yml /etc/
+fi
 echo "To activate, open a new bash instance (executable name is 'workspaces')."
 exec bash
